+++ app/views/vendors/show.html.erb
@@ -2,11 +2,11 @@
   h1 {text-align: center;}
 </style>
 
-<h1><%= @vendor.name %></h1>
+<h1><%= @facade.vendor_name %></h1>
 <br><br>
 
 <p>Contact Info:</p>
-<p>Name: <%= @vendor.contact_name %></p>
-<p>Phone: <%= @vendor.contact_phone %></p>
-<p>Credit Accepted?: <%= @vendor.credit_accepted? %></p>
-<p>Description: <%= @vendor.description %></p>
+<p>Name: <%= @facade.vendor_contact_name %></p>
+<p>Phone: <%= @facade.vendor_contact_phone %></p>
+<p>Credit Accepted?: <%= @facade.vendor_credit_accepted %></p>
+<p>Description: <%= @facade.vendor_description %></p>
